day3 task 1 solution

+++ src/main/kotlin/day2/SolutionTask1.kt
@@ -26,7 +26,7 @@ class SolutionTask1 {
             }) {
             0
         } else {
-            line.integers().first()
+            line.integers().first().toInt()
         }
 
     }
